% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAnswers.R
\name{GetAnswers}
\alias{GetAnswers}
\title{Export data from your CC dashboard}
\usage{
GetAnswers(Username, Password, RequiredQidsVect, AfterDate, BeforeDate,
  MaxSampleSize = 0, Randomize = FALSE, ReturnIntersectedSet = FALSE)
}
\arguments{
\item{Username}{your CC username}

\item{Password}{your CC password}

\item{RequiredQidsVect}{a vector of strings specifying what all qids you want to export}

\item{AfterDate}{String of start date in format - "YYYY-MM-DD"}

\item{BeforeDate}{String of end date in format - "YYYY-MM-DD"}

\item{MaxSampleSize}{specifies what is the maximum sample set you wish to export, by default returns all data}

\item{Randomize}{true if you wish to randomize the data in the given time range and then export. by default set to False}

\item{ReturnIntersectedSet}{set to true if you want to pull the data with an intersection (i.e., AND condition) between the Question ID's specified}
}
\value{
R DataFrame with CC data
}
\description{
This function allows you to export data by specifying what question Ids you want to export
}
\details{
Install the following packages- httr, jsonlite before you use this. Performance depends on how much data is present in the specified time frame.
}
\examples{
GetAnswers("username", "password", c("59df48ea81180318b06d0199", "59df48ea81180318b06d019a"), "2018-01-14", "2018-05-15", 2000, TRUE, TRUE)
}
\keyword{GetAnswers}
