% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetQuestions.R
\name{GetQuestions}
\alias{GetQuestions}
\title{Export Questions present from your CC dashboard}
\usage{
GetQuestions(Username, Password, Questionnaire = NULL)
}
\arguments{
\item{Username}{your CC username}

\item{Password}{your CC password}

\item{Questionnaire}{Questionnaire from which you want to retrieve questions. will retrieve all by default}
}
\value{
R DataFrame with CC data
}
\description{
This function allows you to export all questions or from a select questionnaire
}
\details{
Install the following packages- httr, jsonlite before you use this. Performance depends on how much data is present in the specified time frame.
}
\examples{
GetQuestions("username", "password", "Test Questionnaire")
}
\keyword{GetQuestions}
